Update README.md
This is a test commit to see whether the markdown renders as nicely as we wanted it to be. Do they actually suggest a limit for commits or are they just going to let me type as much as I want. Doesn't seem like they enforce it here... must only be in a different field. Kinda crap. 